scripts/requirements.txt
google-generativeai==0.3.2
notion-client==2.2.1
python-dotenv==1.0.0
PyYAML==6.0.1
